Feat/recurrency

+++ src/controllers/entries.controller.ts
@@ -2,13 +2,27 @@ import dayjs from 'dayjs';
 import { NextFunction, Request, Response } from 'express';
 import { StatusCodes } from 'http-status-codes';
 import prisma from '../lib/prisma';
+import { v4 as uuidv4 } from 'uuid';
 
 interface ITypeEntries {
   [key: string]: string;
 }
-interface IInverseTypeEntries {
-  [key: string]: string;
+interface EntriesSet {
+  userId: string;
+  accountsId: string;
+  categoriesId: string;
+  subCategoriesId: string;
+  title: string;
+  amount: string;
+  typeId: string;
+  realize: boolean;
+  dueDate: string;
+  recurrency: boolean;
+  recurrencyTimes: number;
+  recurrencyIndex: number;
+  recurrencyId: string;
 }
+
 class EntriesController {
   async findAll(req: Request, res: Response, next: NextFunction) {
     const { userId } = req.body;
@@ -41,6 +55,10 @@ class EntriesController {
         accountsId: entrie.accountsId,
         subCategoriesId: entrie.subCategoriesId,
         dueDate: dayjs(entrie.dueDate).format('DD/MM/YYYY'),
+        recurrencyId: entrie.recurrencyId ?? false,
+        recurrency: entrie.recurrency,
+        recurrencyTimes: entrie.recurrencyTimes,
+        recurrencyIndex: entrie.recurrencyIndex,
       };
     });
     res.status(StatusCodes.OK).json(result);
@@ -55,6 +73,9 @@ class EntriesController {
       amount,
       realize,
       dueDate,
+      recurrency,
+      typeRecurrency,
+      recurrencyTimes,
     } = req.body;
     const typeEntries = {
       income: process.env.INCOME_ID || '5745c1c0-053e-4d51-941e-66c7c7bc24f3',
@@ -63,27 +84,63 @@ class EntriesController {
 
     type TTypeEntrie = keyof typeof typeEntries;
     const type: TTypeEntrie = req.body.type;
-
-    const entries = await prisma.entries.create({
-      data: {
-        userId,
-        accountsId,
-        categoriesId,
-        subCategoriesId,
-        title,
-        amount,
-        typeId: typeEntries[type],
-        realize,
-        dueDate: dayjs(dueDate, 'DD/MM/YYYY').format(),
-      },
-    });
-    if (!entries) {
-      return next({
-        status: StatusCodes.BAD_REQUEST,
-        message: 'Não foi possível cadastrar esse lançamento',
+    if (recurrency === true) {
+      const recurrencyId = uuidv4();
+      const recurrencyData = new Set<EntriesSet>();
+      const repeatTimes =
+        recurrencyTimes !== 0 ? recurrencyTimes : process.env.MAX_RECURRENCY;
+      for (let i = 0; i < repeatTimes; i++) {
+        const addInstallments = i;
+        recurrencyData.add({
+          userId,
+          accountsId,
+          categoriesId,
+          subCategoriesId,
+          title,
+          amount,
+          typeId: typeEntries[type],
+          realize: i === 0 ? realize : false,
+          dueDate: dayjs(dueDate, 'DD/MM/YYYY')
+            .add(addInstallments, typeRecurrency)
+            .format(),
+          recurrency,
+          recurrencyId,
+          recurrencyTimes,
+          recurrencyIndex: i,
+        });
+      }
+      const entries = await prisma.entries.createMany({
+        data: Array.from(recurrencyData),
       });
+      if (!entries) {
+        return next({
+          status: StatusCodes.BAD_REQUEST,
+          message: 'Não foi possível cadastrar esse lançamento',
+        });
+      }
+      res.status(StatusCodes.OK).json(entries);
+    } else {
+      const entries = await prisma.entries.create({
+        data: {
+          userId,
+          accountsId,
+          categoriesId,
+          subCategoriesId,
+          title,
+          amount,
+          typeId: typeEntries[type],
+          realize,
+          dueDate: dayjs(dueDate, 'DD/MM/YYYY').format(),
+        },
+      });
+      if (!entries) {
+        return next({
+          status: StatusCodes.BAD_REQUEST,
+          message: 'Não foi possível cadastrar esse lançamento',
+        });
+      }
+      res.status(StatusCodes.OK).json(entries);
     }
-    res.status(StatusCodes.OK).json(entries);
   }
   async realizeEntries(req: Request, res: Response, next: NextFunction) {
     const { id, realize, dueDate } = req.body;
@@ -143,20 +200,67 @@ class EntriesController {
   }
 
   async deleteEntries(req: Request, res: Response, next: NextFunction) {
-    const { id } = req.params;
-    const deleteEntrie = await prisma.entries.delete({
-      where: {
-        id,
-      },
-    });
-
-    if (!deleteEntrie) {
-      return next({
-        status: StatusCodes.BAD_REQUEST,
-        message: 'Não foi possível atualizar o lançamento',
+    const { id, recurrencyId, recurrencyIndex, deleteRecurrency } = req.params;
+    const index = Number(recurrencyIndex);
+    if (deleteRecurrency === 'all') {
+      const deleteEntrie = await prisma.entries.deleteMany({
+        where: {
+          recurrencyId,
+        },
       });
+      if (!deleteEntrie) {
+        return next({
+          status: StatusCodes.BAD_REQUEST,
+          message: 'Não foi possível atualizar o lançamento',
+        });
+      }
+      res.status(StatusCodes.OK).json({ message: 'Lançamento excluído' });
     }
-    res.status(StatusCodes.OK).json({ message: 'Lançamento excluído' });
+    if (deleteRecurrency === 'nexts') {
+      const deleteEntrie = await prisma.entries.deleteMany({
+        where: {
+          recurrencyId,
+          recurrencyIndex: {
+            gte: index,
+          },
+        },
+      });
+      if (!deleteEntrie) {
+        return next({
+          status: StatusCodes.BAD_REQUEST,
+          message: 'Não foi possível atualizar o lançamento',
+        });
+      }
+      res.status(StatusCodes.OK).json({ message: 'Lançamento excluído' });
+    }
+    if (deleteRecurrency === 'unique') {
+      const deleteEntrie = await prisma.entries.delete({
+        where: {
+          id,
+        },
+      });
+      if (!deleteEntrie) {
+        return next({
+          status: StatusCodes.BAD_REQUEST,
+          message: 'Não foi possível atualizar o lançamento',
+        });
+      }
+      res.status(StatusCodes.OK).json({ message: 'Lançamento excluído' });
+    }
+    if (deleteRecurrency === 'none') {
+      const deleteEntrie = await prisma.entries.delete({
+        where: {
+          id,
+        },
+      });
+      if (!deleteEntrie) {
+        return next({
+          status: StatusCodes.BAD_REQUEST,
+          message: 'Não foi possível atualizar o lançamento',
+        });
+      }
+      res.status(StatusCodes.OK).json({ message: 'Lançamento excluído' });
+    }
   }
 }
